Add files via upload
Very Basic GUI created, allows the user to choose between Starting the game and viewing a brief explanation of the games mechanics, should probably be cleaned up to look more appealing. UmiBackGround.jpg is a file I was considering using as the background for the Main Menu, will look into implementing.